Dockerfile: add sanity-check-prepare
To integrate TuxRun with TuxLAVA and add TuxLAVA as dependency, there
is need to build dep and rpm package for the same. To accomodate that
first we need to check for deb and rpm sanity check prepare which
depends on changes included in container. So build the container with
the relevant changes.

Signed-off-by: Alok Ranjan <alok.ranjan@linaro.org>