assets: use tuxlava TEST_DEFINITIONS
It would be better to use tuxlava TEST_DEFINITIONS as it will help in
avoiding future issues due to different version of TEST_DEFINITIONS in
tuxrun and tuxlava. Also it makes to manage and maintain easier as for
any change in TEST_DEFINITION tags, update will be needed at a single
place.

Signed-off-by: Alok Ranjan <alok.ranjan@linaro.org>